Refactor tests from redispy4_plus
package does not support redis below v4.2, therefore there is no need
for these tests to be in a separate file.
