+++ test/test_general.py
@@ -7,3 +7,9 @@ def test_singleton():
     conn2 = conn_generator(dict(), False)
     assert conn1.set('foo', 'bar') is True
     assert conn2.get('foo') == b'bar'
+
+
+def test_asyncioio_is_used():
+    """Redis 4.2+ has support for asyncio and should be preferred over aioredis"""
+    from fakeredis import aioredis
+    assert not hasattr(aioredis, "__version__")
