+++ test/test_list_commands.py
@@ -6,6 +6,7 @@ import redis
 import redis.client
 import redis.client
 
+import testtools
 from testtools import raw_command
 
 
@@ -535,3 +536,66 @@ def test_empty_list(r):
     r.rpush('foo', 'bar')
     r.rpop('foo')
     assert not r.exists('foo')
+
+
+def test_lmove_to_nonexistent_destination(r):
+    r.rpush('foo', 'one')
+    assert r.lmove('foo', 'bar', 'RIGHT', 'LEFT') == b'one'
+    assert r.rpop('bar') == b'one'
+
+
+def test_lmove_expiry(r):
+    r.rpush('foo', 'one')
+    r.rpush('bar', 'two')
+    r.expire('bar', 10)
+    r.lmove('foo', 'bar', 'RIGHT', 'LEFT')
+    assert r.ttl('bar') > 0
+
+
+def test_lmove_wrong_type(r):
+    r.set('foo', 'bar')
+    r.rpush('list', 'element')
+    with pytest.raises(redis.ResponseError):
+        r.lmove('foo', 'list', 'RIGHT', 'LEFT')
+    assert r.get('foo') == b'bar'
+    assert r.lrange('list', 0, -1) == [b'element']
+    with pytest.raises(redis.ResponseError):
+        r.lmove('list', 'foo', 'RIGHT', 'LEFT')
+    assert r.get('foo') == b'bar'
+    assert r.lrange('list', 0, -1) == [b'element']
+
+
+def test_lmove(r):
+    assert r.lmove('foo', 'bar', 'RIGHT', 'LEFT') is None
+    assert r.lpop('bar') is None
+    r.rpush('foo', 'one')
+    r.rpush('foo', 'two')
+    r.rpush('bar', 'one')
+
+    # RPOPLPUSH
+    assert r.lmove('foo', 'bar', 'RIGHT', 'LEFT') == b'two'
+    assert r.lrange('foo', 0, -1) == [b'one']
+    assert r.lrange('bar', 0, -1) == [b'two', b'one']
+    # LPOPRPUSH
+    assert r.lmove('bar', 'bar', 'LEFT', 'RIGHT') == b'two'
+    assert r.lrange('bar', 0, -1) == [b'one', b'two']
+    # RPOPRPUSH
+    r.rpush('foo', 'three')
+    assert r.lmove('foo', 'bar', 'RIGHT', 'RIGHT') == b'three'
+    assert r.lrange('foo', 0, -1) == [b'one']
+    assert r.lrange('bar', 0, -1) == [b'one', b'two', b'three']
+    # LPOPLPUSH
+    assert r.lmove('bar', 'foo', 'LEFT', 'LEFT') == b'one'
+    assert r.lrange('foo', 0, -1) == [b'one', b'one']
+    assert r.lrange('bar', 0, -1) == [b'two', b'three']
+
+    # Catch instances where we store bytes and strings inconsistently
+    # and thus bar = ['two', b'one']
+    assert r.lrem('bar', -1, 'two') == 1
+
+
+@pytest.mark.disconnected
+@testtools.fake_only
+def test_lmove_disconnected_raises_connection_error(r):
+    with pytest.raises(redis.ConnectionError):
+        r.lmove(1, 2, 'LEFT', 'RIGHT')
