+++ test/test_sortedset_commands.py
@@ -58,6 +58,29 @@ def test_zrange_same_score(r):
     assert r.zrange('foo', 2, 3) == [b'two_c', b'two_d']
 
 
+def test_zrange_with_byscore(r):
+    testtools.zadd(r, 'foo', {'zero': 0})
+    testtools.zadd(r, 'foo', {'two': 2})
+    testtools.zadd(r, 'foo', {'two_a_also': 2})
+    testtools.zadd(r, 'foo', {'two_b_also': 2})
+    testtools.zadd(r, 'foo', {'four': 4})
+    assert r.zrange('foo', 1, 3, byscore=True) == [b'two', b'two_a_also', b'two_b_also']
+    assert r.zrange('foo', 2, 3, byscore=True) == [b'two', b'two_a_also', b'two_b_also']
+    assert (
+            r.zrange('foo', 0, 4, byscore=True)
+            == [b'zero', b'two', b'two_a_also', b'two_b_also', b'four']
+    )
+    assert r.zrange('foo', '-inf', 1, byscore=True) == [b'zero']
+    assert (
+            r.zrange('foo', 2, '+inf', byscore=True)
+            == [b'two', b'two_a_also', b'two_b_also', b'four']
+    )
+    assert (
+            r.zrange('foo', '-inf', '+inf', byscore=True)
+            == [b'zero', b'two', b'two_a_also', b'two_b_also', b'four']
+    )
+
+
 def test_zcard(r):
     testtools.zadd(r, 'foo', {'one': 1})
     testtools.zadd(r, 'foo', {'two': 2})
